Update Folder Structuring and Code Testing
