--- Projects/Agents/Resumegents/Backend/tailored_resume_20250111_185013.tex
@@ -1,76 +1,0 @@
-\documentclass[11pt]{article}
-\usepackage[left=0.75in,top=0.6in,right=0.75in,bottom=0.6in]{geometry}
-\usepackage{hyperref}
-\usepackage{fontawesome}
-\usepackage{titlesec}
-\usepackage{enumitem}
-
-\titleformat{\section}{\Large\scshape\raggedright}{}{0em}{}[\titlerule]
-\titlespacing*{\section}{0pt}{12pt}{8pt}
-
-\hypersetup{
-    colorlinks=true,
-    linkcolor=blue,
-    filecolor=magenta,      
-    urlcolor=cyan,
-    pdftitle={Resume - AZEEZ IDRIS}
-}
-
-\begin{document}
-
-\begin{center}
-    {\Huge \textbf{AZEEZ IDRIS}}\\[0.5em]
-    \href{mailto:aidris@iastate.edu}{aidris@iastate.edu} • 
-    \href{tel:+15153570454}{+1 515 357 0454} •
-    Redmond, WA, USA\\
-    \href{https://www.linkedin.com/in/azibit/}{\faLinkedin\ LinkedIn}
-\end{center}
-
-\section{Summary}
-Data Scientist with strong programming skills and experience in developing machine learning solutions. Proficient in Python development with extensive experience in implementing end-to-end solutions.
-
-\section{Technical Skills}
-\textbf{Programming Languages:} Python, Java\\
-\textbf{Frameworks \& Tools:} TensorFlow, PyTorch, Git, REST API\\
-\textbf{Cloud Platforms:} AWS S3, Google Cloud\\
-\textbf{Areas of Expertise:} Natural Language Processing, Machine Learning, Deep Learning
-
-\section{Professional Experience}
-\textbf{Data Scientist II} \hfill Mar 2023 - Present\\
-Microsoft, Redmond, WA
-\begin{itemize}[leftmargin=*]
-    \item Developed Python-based solutions for enhanced user experience and security implementations
-    \item Built and maintained RESTful APIs for AI-driven features including AI Stickers and AI Themes
-    \item Achieved 24\% CTR increase through data-driven optimization in the Personas project
-\end{itemize}
-
-\textbf{Data Scientist} \hfill May 2022 - Aug 2022\\
-Microsoft, Redmond, WA
-\begin{itemize}[leftmargin=*]
-    \item Developed Python-based Natural Language processing pipeline for data collection
-    \item Implemented end-to-end deep-learning models using Python and modern ML frameworks
-\end{itemize}
-
-\textbf{Ph.D. Research} \hfill May 2021 - Aug 2021\\
-IBM, Yorktown Heights, NY
-\begin{itemize}[leftmargin=*]
-    \item Created and implemented machine learning models using Python for job posting classification
-    \item Developed data processing pipelines handling 6000+ technology job postings
-\end{itemize}
-
-\section{Education}
-\textbf{Ph.D. in Computer Science} \hfill Present\\
-Iowa State University \hfill GPA: 3.73\\[0.5em]
-\textbf{MSc in Computer Science} \hfill 2019\\
-Nile University of Nigeria \hfill GPA: 4.73\\[0.5em]
-\textbf{BSc in Computer Science} \hfill 2015\\
-Ahmadu Bello University \hfill GPA: 4.62
-
-\section{Certifications}
-\begin{itemize}[leftmargin=*]
-    \item Oracle Certified Associate, Java SE 7 Programmer
-    \item Introduction To TensorFlow For AI, Machine Learning and Deep Learning
-    \item Machine Learning - Stanford University
-\end{itemize}
-
-\end{document}
