--- Projects/Agents/Resumegents/PROMPTS/ComplexResumeCreator.md
@@ -1,19 +1,0 @@
-Create a complete LaTeX resume based on the provided resume data and any additional instructions that highlights relevant experience for this job. 
-Only use information from the provided resume data and from the additional instructions - do not add or fabricate any new experiences or qualifications that cannot be found in the resume data.
-Match and emphasize experiences from the resume data and the additional instructions that align with the job description requirements.
-
-The LaTeX code should:
-1. Use the article document class
-2. Include necessary packages (geometry, hyperref, etc.)
-3. Have professional formatting
-4. Include all relevant sections based on the provided resume data
-5. Be ready to compile in Overleaf
-6. Ensure you follow any additional formatting or content instructions provided
-
-Important:
-- Only use information from the provided resume data
-- Follow any specific instructions or feedback provided
-- Do not add any experiences or qualifications that are not in the source data
-- Focus on highlighting existing experiences that match the job requirements
-
-Return ONLY the complete LaTeX code.
